Fixing the calculation of the number of days elapsed after the start of the Gregorian Year by taking into account the daylight saving time.
Previously, ,you may have noticed that (with previous version) the
Gregorian Dates 2016-03-26 and 2016-03-27 were both translated into
Bulgarian Date: 4 Твирем, 6784 Бечин (4 Tvirem, 6784 Bechin), while
the correct would be:

Gregorian 2016-03-26 (Saturday)  equivalent to Bulgarian 4 Tvirem
Gregorian 2016-03-27 (Sunday)    equivalent to Bulgarian 5 Tvirem

The reason for that but was the daylignt saving time changes that
apply on the last weekend of March.
https://bg.wikipedia.org/wiki/%D0%9B%D1%8F%D1%82%D0%BD%D0%BE_%D1%87%D0%B0%D1%81%D0%BE%D0%B2%D0%BE_%D0%B2%D1%80%D0%B5%D0%BC%D0%B5

The issue comes from the fact that the difference between

2016-01-01 00h 00m    and 2016-03-26 00h 00m    is exactly 85 days (24 hours each)

but the difference between

2016-01-01 00h 00m    and 2016-03-27 00h 00m    is 85 days (24 hours each) + 23 hours
(because we have moved the clock one hour ahead based on the daylignt saving time)

Thus the calculation d1.getTime() - d2.getTime() / DAY
returns 85.9... for (2016-03-27) and then

Math.floor(85.9)

gives us 85 days, instead of the expected 86 days.

--------------------------------------------------
TODO:

Even after that fix, the calendar is not absolutely correct, because
it assumes that a leap year in a Gregorian Calendar is a leap year in the
Bulgarian calendar.

That assumption seems to be wrong according to some researcherrs.
They actually claim that the Bulgarian calendar is more precise than the
Gregorian calendar. Some claim it is even the most precise Calendar ever
invented.

That precision is based on a more precise model of calculating leap years
which differs from that in the Gregorian Calendar.

For reference please see this article:
http://pencho.my.contact.bg/start/razni/journey/explain.htm

which is also part of the used literature on the site:
http://calendar.samoistina.com/

Please note the last paragraph regarding correction of the calendar
While Gregorian calendar has corrections on 4 and 400 years,
the Bulgarian has corrections on 4, 420, 1680, 20160, 80640 and 10080000
years.